requirements.txt
temporalio==1.9.0
requests==2.28.1
beautifulsoup4==4.11.1
openai
python-dotenv==0.21.0
